Does not include empty string parameters in SQL in statements
Fixes issue #68 by ensuring that the list passed to GORM for in
statements does not include empty strings, however once the list has
more than one item there is a performance degradation again. Also limits
the count query to 300k for speed, it was deemed unlikely that users
will want to paginate up to or beyond 300k matching records. This figure
is not magic, it is also the upper limit for KML exports so the two
concepts are related.