Removed the vagrant/guest stuff from the restore script. Tidied up the backup/restore doc a little.
It's now assumed that the VM exists and has a provisioned DB, and that the script will run from within the VM (rather than via `vagrant ssh`).

The main reason for doing this is to enable @xhoenner to execute the script in an already running VM. This is because he doesn't have access to chef repo (therefore can't run the `vagrant up` etc commands).