[Work in progress] Unit tests for protected species feature
A work-in-progress commit for the protected species feature.
I am currently working to get the KML export unit-tested.
We also need to ensure that the 'associated entities' are filtered as required by the spec.
We should also update the spec document and test code to use a term that is more clear than 'unembargoed' (e.g. 'public', 'no protection', etc.)
The way the protected species tests are named and organised needs to be reviewed. I reckon we can make then tidier.

There is a bunch of work on this branch still to be tidied-up, too. There are println()s, to-do comments, etc.