Added views which mimic the current (in)valid_detection tables
The main difference being that validity is derived (from the existence or not of associated metadata).

Another small difference, which can possibly be addressed, is that the derived validity in this change doesn't differentiate between no associated deployment and no associated recovery.  This may or may not be worth solving.

This change puts the database in a position where we can start running benchmarks and hopefully be mostly compatible with the current code base (except for the view names - this can be easily changed later).