Use liquibase to configure test database
This is to simplify test setup, e.g. no need to "manually" create views in tests. It should make it easier to fix up tests in the "detection validity" work that is in progress.

Deleted a couple of tests which were broken and/or about to become redundant when the above is complete.