Fixes an exception when viewing individual detection records
In practise, there is a 1:many relationship between detections and surgeries (this is because of tag ping code recycling, albeit a rare occurance).

Previously, this was manifest in the app with one record per detection on the list view, and then in `detection show`, the "many" surgeries being shown.

Now, multiple records will be listed where there are many surgeries, with only the one surgery being shown in the "show" view.

This has changed because of the changes to the underlying database view.