Device status now represented as an enum type
... rather than a database type.
This is in order to simplify code surrounding device statuses, e.g. the bug fixing which is currently happening around overlapping deployment invalidation.

It comes at the cost of it now being harder to add/edit device statuses (i.e. it requires a code change), but the set of statuses has been stable for quite some time, so we can probably live with it.