Search indexing is now done on main thread
This is to avoid getting spurious exception sometimes when shutting down the app (because forked indexing hasn't finished).  This became especially apparent when running database migration commands such as `update` and `rollback`.

It means that start up takes ~20s longer, but this is ok.

Also, deleted a lot of the documentation from the file - if we need to know what the configuration options are, we can look it up.