Addresses #248
The inclusion of the transmitter name column is somewhat misleading (as it depends on what users have entered in their own VUE database, rather than what's in the AATAMS DB).  It also constitutes a potential security risk w.r.t. to species information, as some users are naming their tags after the species to which they're attached, however this information is not subject to any restriction/embargo (i.e. it's always publicly visible).

For these reasons, this column has been removed.