Fix for #258
Previously, the `detection_count_per_station` view, and associated materialised view, was being used for much more than just getting the number of detections at each station.  For example, it was used in the past to serve up some WMS/WFS layers.

This is no longer the case, the view can be greatly simplified and therefore runs in much less time (~15 minutes on production).