Fix issue running PersonControllerIntegrationTests on xenial/postgres 9.6
Audit log plugin was trying to write an audit log entry when creating the organisation, calling Person.findByName to find the user which was
flushing the address changes to the database a second time causing a duplicate key error.  This doesn't reflect how the user will
actually be found (from the session) so changed this to return the user setup in the test setup method which is closer to how this works.